+++ config/initializers/content_security_policy.rb
@@ -12,6 +12,7 @@ Rails.application.config.content_security_policy do |policy|
   # (https://fonts.gstatic.com at time of writing, but I see nothing to
   # suggest we can hardcode this host)
   policy.font_src    :self, :https
+  policy.script_src  :self, :unsafe_inline
 end
 
 # If you are using UJS then enable automatic nonce generation