+++ spec/features/view_supplier_details_spec.rb
@@ -7,20 +7,20 @@ RSpec.feature 'Viewing a supplier' do
 
   before { sign_in_as_admin }
 
-  scenario 'shows list of frameworks the supplier has an agreement for' do
+  scenario 'shows paginated list of frameworks the supplier has an agreement for' do
     visit admin_supplier_path(supplier)
     expect(page).to have_content 'Test Supplier Ltd'
     expect(page).to have_content 'RM0000 Test Framework'
-    # expect(page).to have_content 'Displaying 1 framework'
+    expect(page).to have_content 'Displaying 1 framework'
   end
 
-  scenario 'shows list of the supplier’s tasks' do
+  scenario 'shows paginated list of the supplier’s tasks' do
     FactoryBot.create(:task, period_month: 12, period_year: 2018, supplier: supplier, framework: framework)
 
     visit admin_supplier_path(supplier)
     expect(page).to have_content 'December 2018'
     expect(page).to have_content 'Unstarted'
-    # expect(page).to have_content 'Displaying 1 task'
+    expect(page).to have_content 'Displaying 1 task'
   end
 
   scenario 'includes the details of a task’s submissions' do
@@ -52,10 +52,45 @@ RSpec.feature 'Viewing a supplier' do
     expect(page.body).to include File.open(Rails.root.join('spec', 'fixtures', 'test.xlsx'), 'r:ASCII-8BIT', &:read)
   end
 
-  scenario 'shows list of the users linked to the supplier' do
+  scenario 'shows paginated list of the users linked to the supplier' do
     FactoryBot.create(:user, suppliers: [supplier])
 
     visit admin_supplier_path(supplier)
     expect(page).to have_content 'Active?'
+    expect(page).to have_content 'Displaying 1 user'
+  end
+
+  scenario 'allows users to be filtered by status' do
+    FactoryBot.create(:user, name: 'Active User', suppliers: [supplier])
+    FactoryBot.create(:user, name: 'Inactive User', auth_id: nil, suppliers: [supplier])
+
+    visit admin_supplier_path(supplier)
+
+    expect(page).to have_content 'Active User'
+    expect(page).to have_content 'Inactive User'
+    expect(page).to have_content 'Filter Status'
+
+    page.check('status[]', match: :first)
+    find('#status-filter-submit').click
+
+    expect(page).to have_content 'Active User'
+    expect(page).not_to have_content 'Inactive User'
+  end
+
+  scenario 'each section\'s pagination works independently' do
+    period_month = 0
+    12.times do
+      FactoryBot.create(:task, period_month: period_month += 1, period_year: 2018, supplier: supplier,
+framework: framework)
+      FactoryBot.create(:user, suppliers: [supplier])
+    end
+    FactoryBot.create(:task, period_month: 12, period_year: 2017, supplier: supplier, framework: framework)
+    FactoryBot.create(:user, suppliers: [supplier])
+
+    visit admin_supplier_path(supplier)
+    click_link('Next »', match: :first)
+    expect(page).to have_content 'Displaying task 13 - 13 of 13 in total'
+    expect(page).to have_content 'Displaying users 1 - 12 of 13 in total'
+    expect(page).to have_content 'Displaying 1 framework'
   end
 end